backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
slither-analyzer==0.10.0
groq==0.4.1
python-dotenv==1.0.0